# (PC-21991)[API] script: to fill product.ean and offer.ean