Lien vers le ticket Jira : https://passculture.atlassian.net/browse/PC-21991

[x] Write isbn in isbn and ean when creating a product or an offer
[] Duplicate isbn in ean (We are here !)
[] Delete isbn keys and values from all products / offers

Script to copy every isbn in ean. This information is stored in a json field inside Offer or Product table. 
--- Old performance ---
ETA : 
For 100 000 with batches of 1000 : 
db.session.bulk_update_mappings -> 155s
db.session.add(product) V1 -> 86s
db.session.add(product) V2 -> 87s

For 100 000 with batches of 10000 : 
db.session.bulk_update_mappings -> 146s
db.session.add(product) V1 -> 78.3s
db.session.add(product) V2 -> 78.9s

There is around 3 733 340 products and more than 79 448 037 offers. 
The script should take ~0.8h for products and ~17h for offers choosing 80s for 100 000 objects. 

I think i will use the db.session.add V2. 
I plan to have a single script to handle Product or Offer (with a param) and to be able to start from a specific id if it stops during execution.
--- Old performance ----

Thanks to some advices we were able to be more efficient . This version took 14s for 100 000 objects. 
It should take less than 10 min for products and 3h for offers